% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiv_incidence.R
\docType{data}
\name{hiv_incidence}
\alias{hiv_incidence}
\title{HIV Incidence Data}
\format{
A data frame with 11 rows and 3 columns:
\describe{
  \item{country}{Factor. The name of the country. For this dataset, all entries are "Angola".}
  \item{year}{Integer. The year of the observation, ranging from 1990 to 2000.}
  \item{new_cases}{Integer. The number of new HIV cases reported in that year.}
}
}
\source{
Data collected from [source of your data].
}
\usage{
hiv_incidence
}
\description{
This dataset contains information on the number of new HIV cases reported annually in Angola from 1990 to 2000.
}
\keyword{datasets}
